chore: refactor core router module
