+++ packages/desktop/App.svelte
@@ -15,7 +15,7 @@
     import { appSettings, appStage, AppStage, appVersionDetails, initAppSettings } from '@core/app'
     import { Electron } from '@lib/electron'
     import { showAppNotification } from '@auxiliary/notification'
-    import { openPopup, popupState } from '@auxiliary/popup'
+    import { closePopup, openPopup, popupState } from '@auxiliary/popup'
     import { Dashboard, LoginRouter, OnboardingRouter, Settings, Splash } from 'shared/routes'
     import { onDestroy, onMount } from 'svelte'
     import { getLocalisedMenuItems } from './lib/helpers'
@@ -89,6 +89,7 @@
         })
         Electron.onEvent('menu-navigate-settings', () => {
             if ($loggedIn) {
+                closePopup()
                 openSettings()
             } else {
                 settings = true
