+++ packages/mobile/lib/contexts/dashboard/constants/dashboard-tab-component.constant.ts
@@ -1,5 +1,5 @@
 import { ActivityTab, TokensTab } from '../../../../views/dashboard/tabs'
-import { DashboardTab } from '../enums/dashboard-tab.enum'
+import { DashboardTab } from '../enums'
 
 export const DASHBOARD_TAB_COMPONENT = {
     [DashboardTab.Tokens]: TokensTab,
