+++ packages/mobile/lib/contexts/dashboard/constants/initial-active-dashboard-tab.constant.ts
@@ -1,9 +1,9 @@
 import { DashboardTab } from '../enums'
 import features from '../../../../features/features'
 
-export const INITIAL_ACTIVE_DASHBOARD_TAB: DashboardTab | null = getInitialActiveTab()
+export const INITIAL_ACTIVE_DASHBOARD_TAB: DashboardTab | null = getInitialActiveDashboardTab()
 
-function getInitialActiveTab(): DashboardTab | null {
+function getInitialActiveDashboardTab(): DashboardTab | null {
     if (features?.dashboard?.tokens?.enabled) {
         return DashboardTab.Tokens
     } else if (features?.dashboard?.activity?.enabled) {
