+++ packages/shared/lib/core/router/actions/resetWalletRoute.ts
@@ -1,0 +1,7 @@
+import { get } from 'svelte/store'
+
+import { dashboardRouter } from '../routers'
+
+export function resetWalletRoute(): void {
+    get(dashboardRouter).reset()
+}
