+++ packages/shared/lib/core/router/index.ts
@@ -1,12 +1,7 @@
+export * from './actions'
 export * from './classes'
 export * from './enums'
 export * from './interfaces'
+export * from './routers'
 export * from './subrouters'
 export * from './types'
-
-export * from './app-router'
-export * from './collectibles-router'
-export * from './dashboard-router'
-export * from './helper'
-export * from './onboarding-router'
-export * from './settings-router'
