+++ packages/shared/lib/core/router/interfaces/index.ts
@@ -1,1 +1,2 @@
 export * from './router.interface'
+export * from './router-event.interface'
