+++ packages/shared/lib/core/router/routers/app-router.ts
@@ -2,10 +2,10 @@ import { get, writable } from 'svelte/store'
 
 import { profiles } from '@core/profile'
 
-import { Router } from './classes'
-import { AppRoute, LoginRoute } from './enums'
-import { loginRoute } from './subrouters'
-import { IRouterEvent } from './types'
+import { Router } from '../classes'
+import { AppRoute, LoginRoute } from '../enums'
+import { IRouterEvent } from '../interfaces'
+import { loginRoute } from '../subrouters'
 
 export const appRoute = writable<AppRoute>(null)
 export const appRouter = writable<AppRouter>(null)
