+++ packages/shared/lib/core/router/routers/collectibles-router.ts
@@ -1,7 +1,7 @@
 import { writable } from 'svelte/store'
 
-import { Router } from './classes'
-import { CollectiblesRoute } from './enums'
+import { Router } from '../classes'
+import { CollectiblesRoute } from '../enums'
 
 export const collectiblesRouter = writable<CollectiblesRouter>(null)
 export const collectiblesRoute = writable<CollectiblesRoute>(null)
