+++ packages/shared/lib/core/router/routers/index.ts
@@ -1,0 +1,5 @@
+export * from './app-router'
+export * from './collectibles-router'
+export * from './dashboard-router'
+export * from './onboarding-router'
+export * from './settings-router'
