+++ packages/shared/lib/core/router/subrouters/onboarding/app-setup-router.ts
@@ -1,8 +1,8 @@
 import { get, writable } from 'svelte/store'
 
 import { AppSetupRoute } from '../../enums'
-import { onboardingRouter } from '../../onboarding-router'
 import { Subrouter } from '../../classes'
+import { onboardingRouter } from '../../routers'
 
 export const appSetupRoute = writable<AppSetupRoute>(null)
 export const appSetupRouter = writable<AppSetupRouter>(null)
