+++ packages/shared/lib/core/router/subrouters/onboarding/ledger-setup-router.ts
@@ -1,8 +1,8 @@
 import { get, writable } from 'svelte/store'
 
-import { LedgerSetupRoute } from '../../enums'
-import { onboardingRouter } from '../../onboarding-router'
 import { Subrouter } from '../../classes'
+import { LedgerSetupRoute } from '../../enums'
+import { onboardingRouter } from '../../routers'
 
 export const ledgerSetupRoute = writable<LedgerSetupRoute>(null)
 export const ledgerSetupRouter = writable<LedgerSetupRouter>(null)
