+++ packages/shared/lib/core/router/subrouters/onboarding/network-setup-router.ts
@@ -3,9 +3,9 @@ import { get, writable } from 'svelte/store'
 import { onboardingProfile } from '@contexts/onboarding'
 import { NetworkType } from '@core/network'
 
-import { NetworkSetupRoute } from '../../enums'
-import { onboardingRouter } from '../../onboarding-router'
 import { Subrouter } from '../../classes'
+import { NetworkSetupRoute } from '../../enums'
+import { onboardingRouter } from '../../routers'
 
 export const networkSetupRoute = writable<NetworkSetupRoute>(null)
 export const networkSetupRouter = writable<NetworkSetupRouter>(null)
