+++ packages/shared/lib/core/router/subrouters/onboarding/profile-recovery-router.ts
@@ -2,9 +2,9 @@ import { get, writable } from 'svelte/store'
 
 import { onboardingProfile, ProfileRecoveryType } from '@contexts/onboarding'
 
-import { ProfileRecoveryRoute } from '../../enums'
-import { onboardingRouter } from '../../onboarding-router'
 import { Subrouter } from '../../classes'
+import { ProfileRecoveryRoute } from '../../enums'
+import { onboardingRouter } from '../../routers'
 
 export const profileRecoveryRoute = writable<ProfileRecoveryRoute>(null)
 export const profileRecoveryRouter = writable<ProfileRecoveryRouter>(null)
