+++ packages/shared/lib/core/router/subrouters/onboarding/storage-protection-setup-router.ts
@@ -2,9 +2,9 @@ import { get, writable } from 'svelte/store'
 
 import { onboardingProfile, ProfileProtectionType } from '@contexts/onboarding'
 
-import { onboardingRouter } from '../../onboarding-router'
-import { StorageProtectionSetupRoute } from '../../enums'
 import { Subrouter } from '../../classes'
+import { StorageProtectionSetupRoute } from '../../enums'
+import { onboardingRouter } from '../../routers'
 
 export const storageProtectionSetupRoute = writable<StorageProtectionSetupRoute>(null)
 export const storageProtectionSetupRouter = writable<StorageProtectionSetupRouter>(null)
