+++ packages/shared/lib/core/router/types/index.ts
@@ -1,2 +1,1 @@
-export * from './router-event'
 export * from './sidebar'
